Many optimizations to the code
Most of the changes I've done boil down to adding the line gml_pragma("forceinline") to appropriate scripts. It essentially tells the compiler to compile that whole script inline, which increases the performance ever so slightly. That being said, when misused it increases the final executable size considerably, so it should only be used with scripts that are either only single use (like initialization scripts) or scripts that only consist of a single line or very basic logic.

One other major change I did was to change how the whole "pre-freeze" thing was handled. Saving the application surface as a sprite then loading it as a sprite is INCREDIBLY inefficient and slow, so I replaced it with a simple surface.